Create `activityQuadTreeBounds` in `BeamMobsim` and use it to expand `envelopeInUTM`
